requirements.txt
Werkzeug==1.0.0
Flask==1.1.1
requests==2.23.0
python-dateutil==2.8.1
python-dotenv==0.12.0
blinker==1.4
scikit-learn>=0.20.3
numpy>=1.13.3
structlog>=19.1.0
shapely>=1.6.4
torch
torchvision
transformers
matplotlib>=3.0.3
uWSGI==2.0.18
black>=19.3b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docutils==0.15.2
joblib==0.14.1
urllib3==1.25.8